requirements.txt
fastapi
gunicorn
uvicorn
transformers
torch
sentence-transformers
accelerate
bitsandbytes
peft
datasets
google-generativeai
python-dotenv